+++ resources/views/components/bar-stacked.blade.php
@@ -1,0 +1,5 @@
+@props(['series' => [], 'categories' => [], 'legend' => true])
+
+<x-chart::base component="bar-stacked" :option="\WireCharts\Support\Option::barStacked([
+    'series' => $series, 'categories' => $categories, 'legend' => $legend,
+])" {{ $attributes }} />