+++ resources/views/components/column-percent.blade.php
@@ -1,0 +1,5 @@
+@props(['series' => [], 'categories' => [], 'legend' => true])
+
+<x-chart::base component="column-percent" :option="\WireCharts\Support\Option::columnPercent([
+    'series' => $series, 'categories' => $categories, 'legend' => $legend,
+])" {{ $attributes }} />