+++ resources/views/components/histogram.blade.php
@@ -1,0 +1,5 @@
+@props(['data' => [], 'bins' => 10])
+
+<x-chart::base component="histogram" :option="\WireCharts\Support\Option::histogram([
+    'data' => $data, 'bins' => $bins,
+])" {{ $attributes }} />