+++ resources/views/components/lollipop.blade.php
@@ -1,0 +1,5 @@
+@props(['series' => [], 'categories' => []])
+
+<x-chart::base component="lollipop" :option="\WireCharts\Support\Option::lollipop([
+    'series' => $series, 'categories' => $categories,
+])" {{ $attributes }} />