+++ resources/views/components/pareto.blade.php
@@ -1,0 +1,5 @@
+@props(['categories' => [], 'data' => [], 'name' => 'Count'])
+
+<x-chart::base component="pareto" :option="\WireCharts\Support\Option::pareto([
+    'categories' => $categories, 'data' => $data, 'name' => $name,
+])" {{ $attributes }} />